tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.clip.intersect.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.clip.intersect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.clip.intersect</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  ctx.beginPath();
  ctx.rect(0, 0, 50, 50);
  ctx.clip();
  ctx.beginPath();
  ctx.rect(50, 0, 50, 50)
  ctx.clip();

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
